Game Sales vs Console Launch Date
Graph that attempts to show a correlation between the sales of a game
and how long after the console's release date that the game launched.

The top of the axis is the minimum (sales and difference in dates), and
the bottom is the maximum.

Users can press [ or ] to cycle between the console generations.

Sales and Release Information obtained from Wikipedia
and vgchartz.com